Refactor logging and config functions
* set propagate = False on package logger to prevent duplicate logs
* moved logging functions to utils/logging.py
* added utils/config_utils.py
