+++ src/ng_model_gym/api.py
@@ -1,340 +1,26 @@
 # SPDX-FileCopyrightText: <text>Copyright 2024-2025 Arm Limited and/or
 # its affiliates <open-source-office@arm.com></text>
 # SPDX-License-Identifier: Apache-2.0
-import json
 import logging
-import re
-import shutil
-import sys
-import warnings
 from datetime import datetime
-from importlib.resources import files
 from pathlib import Path
-from typing import Callable, List, Tuple, Union
+from typing import Callable, Union
 
 import torch
-from pydantic import ValidationError
-from rich import print_json
-from rich.console import Console
-from rich.table import Column, Table
 from torch.profiler import schedule
 
 from ng_model_gym.nss.evaluators import ModelEvaluator
 from ng_model_gym.trainer import Trainer
 from ng_model_gym.utils.checkpoint_utils import load_checkpoint
-from ng_model_gym.utils.config_model import ConfigModel, OutputDirModel
+from ng_model_gym.utils.config_model import ConfigModel
 from ng_model_gym.utils.general_utils import create_directory
 from ng_model_gym.utils.gpu_log_decorator import gpu_log_decorator
-from ng_model_gym.utils.json_reader import read_json_file
-from ng_model_gym.utils.logging import setup_logging
+from ng_model_gym.utils.logging import log_gpu_torch
 from ng_model_gym.utils.memory_log_decorator import memory_log_decorator
 from ng_model_gym.utils.time_decorator import time_decorator
 from ng_model_gym.utils.types import ExportType, ProfilerType, TrainEvalMode
 
 logger = logging.getLogger(__name__)
-
-DEFAULT_PATH = "ng_model_gym.nss.configs"
-
-
-def generate_config_file(save_dir: Union[str, Path, None] = None) -> Tuple[Path, Path]:
-    """
-    Generate a JSON configuration template and its schema file. This is used to configure training.
-
-    Args:
-        save_dir (Union[str, Path, None]): Directory to save config and schema. If None,
-         uses current directory.
-
-    Returns:
-        Tuple[Path, Path]: Paths to the generated configuration JSON and schema JSON files.
-
-    Example:
-        >>> config_output_path, schema_path = generate_config_file(Path('./output'))
-    """
-
-    if save_dir:
-        output_dir = Path(save_dir)
-        if not output_dir.exists() or not output_dir.is_dir():
-            raise FileNotFoundError(f"Provided save_dir '{output_dir}' does not exist.")
-    else:
-        output_dir = Path(".")
-
-    placeholders = {
-        "dataset.path.train": "<PATH/TO/TRAIN_DATA_DIR>",
-        "dataset.path.validation": "<PATH/TO/VALIDATION_DATA_DIR>",
-        "dataset.path.test": "<PATH/TO/TEST_DATA_DIR>",
-        "train.fp32.checkpoints.dir": "<OUTPUT/PATH/FOR/CHECKPOINTS_DIR>",
-        "train.qat.checkpoints.dir": "<OUTPUT/PATH/FOR/CHECKPOINTS_DIR>",
-    }
-
-    # Load internal default config
-    default_config_path = files(DEFAULT_PATH) / "default.json"
-    default_config: dict = read_json_file(default_config_path)
-
-    for json_loc, placeholder_text in placeholders.items():
-        dict_key_path: List[str] = json_loc.split(".")
-        config_section = default_config
-
-        # Iterate over dict keys e.g dict_key_path = ["dataset", "path", "train"]
-        for dict_key in dict_key_path[:-1]:
-            # Check the json path provided in placeholders is valid
-            if dict_key not in config_section or not isinstance(
-                config_section[dict_key], dict
-            ):
-                raise KeyError(
-                    f"Invalid config path {json_loc} to replace with a placeholder"
-                )
-            # Access sub dict e.g config_section["dataset"]
-            config_section = config_section[dict_key]
-
-        final_key = dict_key_path[-1]
-        if final_key not in config_section:
-            raise KeyError(
-                f"Invalid config path {json_loc} to replace with a placeholder"
-            )
-
-        config_section[final_key] = placeholder_text
-
-    file_name = "config"
-    suffix = ".json"
-    config_output_path = output_dir / f"{file_name}{suffix}"
-
-    # If a file already exists "config.json", create config_1.json etc
-    if config_output_path.exists():
-        count = 1
-        file_path = output_dir / f"{file_name}_{count}{suffix}"
-        while file_path.exists():
-            count += 1
-            file_path = output_dir / f"{file_name}_{count}{suffix}"
-        config_output_path = file_path
-
-    # Write the config file
-    with open(config_output_path, "w", encoding="utf-8") as f:
-        json.dump(default_config, f, indent=4)
-
-    # Copy schema_config.json as well
-    schema_config_path = files(DEFAULT_PATH) / "schema_config.json"
-    shutil.copy(src=schema_config_path, dst=output_dir)
-
-    schema_path = output_dir / "schema_config.json"
-    return config_output_path, schema_path
-
-
-def load_config_file(user_config_path: Path) -> ConfigModel:
-    """
-    Load and validate a JSON configuration file. The created config object is used as an argument
-    to other API functions.
-
-    Note: SystemExit - if validation fails, after printing errors, exits with status code 1
-
-    Args:
-        user_config_path (Path): Path to a user-provided JSON config file
-
-    Returns:
-        ConfigModel: Parsed and validated configuration model.
-
-    Example:
-        >>> params = load_config_file(Path("config.json"))
-    """
-
-    user_config: dict = read_json_file(user_config_path)
-
-    try:
-        return ConfigModel.model_validate(user_config)
-    except ValidationError as e:
-        validation_errors = e.errors()
-
-        # Set up logging for validating config file, logging to file only
-        # Try to get the user's specified output_dir, fall back to ./output if not valid
-        output_dir = user_config.get("output", {}).get("dir")
-
-        try:
-            output_dir = OutputDirModel(dir=output_dir).dir
-
-        except ValidationError:
-            output_dir = Path("./output")
-
-        setup_logging(
-            logger_name=f"{__name__}.config_validation",
-            log_level=logging.ERROR,
-            output_dir=output_dir,
-            stdout=False,
-        )
-
-        config_logger = logging.getLogger(f"{__name__}.config_validation")
-        config_logger.propagate = False
-
-        # Format validation errors into a table
-        table = Table(
-            Column(
-                header="Validation Issue",
-                justify="center",
-                style="red",
-                no_wrap=False,
-                overflow="fold",
-            ),
-            Column(header="Details", justify="center", no_wrap=False, overflow="fold"),
-            Column(
-                header="Your input",
-                justify="center",
-                style="blue",
-                no_wrap=False,
-                overflow="fold",
-            ),
-            Column(
-                header="Location in JSON",
-                justify="center",
-                style="green",
-                no_wrap=False,
-                overflow="fold",
-            ),
-            show_lines=True,
-            expand=True,
-        )
-
-        config_logger.error("Config validation errors:")
-
-        # Create table rows sorted by type e.g. missing, int_parsing
-        for error_ctx in sorted(validation_errors, key=lambda e: e["type"]):
-            val_type = error_ctx["type"]
-            message = error_ctx["msg"]
-
-            user_input = "" if val_type == "missing" else str(error_ctx["input"])
-
-            # Location is a tuple, turn into list of strings and use dot instead of space
-            location_in_json = ".".join(map(str, [*error_ctx["loc"]]))
-
-            table.add_row(val_type, message, user_input, location_in_json)
-
-            # Log errors
-            config_logger.error(
-                f"[{val_type}] {message} (input={user_input}, location={location_in_json})"
-            )
-
-        # Instantiate Rich console and print table
-        console = Console()
-        console.print(table)
-
-        num_errors = len(validation_errors)
-
-        error_str = (
-            f"\n[bold]Configuration has [red]{num_errors}[/red] "
-            f"{'issues' if num_errors > 1 else 'issue'}[/bold]"
-        )
-
-        console.print(error_str)
-
-        error_str = re.sub(r"\[/?[^\]]+\]", "", error_str).lstrip("\n")
-        config_logger.error(error_str)
-
-        sys.exit(1)
-
-
-def print_config_options() -> None:
-    """
-    Print the JSON configuration schema, listing each parameter with its type and description
-
-    Example:
-        >>> print_config_options()
-    """
-    schema_config_path = files(DEFAULT_PATH) / "schema_config.json"
-    with schema_config_path.open("r", encoding="utf-8") as fp:
-        data = json.load(fp)
-
-    print_json(json.dumps(data))
-
-
-def logging_config(
-    params: ConfigModel,
-    logger_name: str = "ng_model_gym",
-    log_level: int = logging.INFO,
-) -> None:
-    """
-    Configure ng_model_gym package logging to file and console.
-
-     Args:
-         params (ConfigModel): Configuration model containing `output.dir`.
-         logger_name (str): Root logger name (e.g., 'ng_model_gym'). Defaults to empty.
-         log_level (int): Logging level (e.g logging.INFO, logging.ERROR). Defaults to logging.INFO.
-
-     Example:
-         >>> logging_config(params, "ng_model_gym", logging.INFO)
-    """
-    # Create output directory if it doesn't exist.
-    create_directory(params.output.dir)
-
-    # Setup general logger which will be used throughout all modules.
-    # The root name is used to allow nested modules to pick up the same logger using __name__.
-    setup_logging(logger_name, log_level, params.output.dir)
-
-    if log_level != logging.DEBUG:
-        # ignore the pretrained deprecation warning in torchvision.models._utils
-        warnings.filterwarnings(
-            "ignore",
-            message=r"The parameter 'pretrained' is deprecated.*",
-            category=UserWarning,
-            module=r"torchvision\.models\._utils",
-        )
-
-        # ignore the weights deprecation warning in torchvision.models._utils
-        warnings.filterwarnings(
-            "ignore",
-            message=r"Arguments other than a weight enum or `None` for 'weights' are deprecated.*",
-            category=UserWarning,
-            module=r"torchvision\.models\._utils",
-        )
-
-        warnings.filterwarnings(
-            "ignore",
-            message=r"TORCH_CUDA_ARCH_LIST is not set, all archs for visible cards are included.*",
-            category=UserWarning,
-            module=r"torch\.utils\.cpp_extension",
-        )
-
-        warnings.filterwarnings(
-            "ignore",
-            message=r"pkg_resources is deprecated as an API.*",
-            category=UserWarning,
-            module=r"slangtorch\.slangtorch",
-        )
-
-        warnings.filterwarnings(
-            "ignore",
-            message=r"pkg_resources is deprecated as an API.*",
-            category=UserWarning,
-            module=r"executorch\.exir\.dialects\.edge\._ops",
-        )
-
-        warnings.filterwarnings(
-            "ignore",
-            message=(
-                r"To copy construct from a tensor, it is recommended to use "
-                r"sourceTensor\.detach\(\)\.clone\(\)*"
-            ),
-            category=UserWarning,
-            module=r"executorch\.backends\.arm\.quantizer\.quantization_config",
-        )
-
-        warnings.filterwarnings(
-            "ignore",
-            message=r"erase_node*",
-            category=UserWarning,
-            module=r"torch\.fx\.graph",
-        )
-
-    if log_level == logging.DEBUG:
-        # Setup PyTorch logger, which will write to the same output file.
-        setup_logging("torch", logging.INFO, params.output.dir)
-
-
-def log_gpu_torch():
-    """Log information about GPU used in training."""
-    if torch.cuda.is_available():
-        num_gpus = torch.cuda.device_count()
-        logger.info(f"GPUs available: {num_gpus}")
-        for i in range(num_gpus):
-            logger.info(f"Using GPU {i}: {torch.cuda.get_device_name(i)}")
-    else:
-        logger.info("No GPU available. Using CPU for training.")
 
 
 def _trace_profiler_wrapper(func: Callable, *args, trace_output_dir: Path):
@@ -418,6 +104,7 @@ def do_training(
         )
 
     params.model_train_eval_mode = training_mode
+
     log_gpu_torch()
 
     trained_model = Trainer(params)
@@ -531,9 +218,6 @@ def do_export(
         ...     export_type=ExportType.FP32,
         ...)
     """
-
-    # Fix for torchao adding BasicHandler if none was present on root
-    logging.getLogger().addHandler(logging.NullHandler())
     from ng_model_gym.utils.export_utils import (  # pylint: disable=import-outside-toplevel
         executorch_vgf_export,
     )
