+++ tests/nss/integration/test_api.py
@@ -5,13 +5,9 @@
 import unittest
 from pathlib import Path
 
-from ng_model_gym.api import (
-    do_evaluate,
-    do_export,
-    do_training,
-    load_config_file,
-    logging_config,
-)
+from ng_model_gym.api import do_evaluate, do_export, do_training
+from ng_model_gym.utils.config_utils import load_config_file
+from ng_model_gym.utils.logging import logging_config
 from ng_model_gym.utils.types import ExportType, ProfilerType, TrainEvalMode
 from tests.nss.integration.base_integration import BaseIntegrationTest
